corpus scoring implemented

+++ src/main/java/jevm/nutshell/engine/TextAnalyzer.java
@@ -6,7 +6,7 @@ import java.util.Map;
 public interface  TextAnalyzer {
 
     public Map<String, Double> getKeywords(int n, String strategy) ;
-    public int scoreSentence(String sentence);
+    public double scoreSentence(String sentence);
     public int scoreWord(String word);
